把武松和157的代码合并

+++ server/src/main/java/com/studyroom/dto/CancelDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 import javax.validation.constraints.NotNull;
 
 /**
- * 取消预约请求DTO
+ * 取消预约请求DTO - 对齐数据库表结构
  * 负责人：严雯琪
  */
 @Data
@@ -16,7 +16,7 @@ public class CancelDTO {
 
     @ApiModelProperty(value = "预约ID", required = true)
     @NotNull(message = "预约ID不能为空")
-    private Long reservationId;
+    private Integer reservationId;
 
     @ApiModelProperty("取消原因")
     private String reason;
