+++ server/src/main/java/com/studyroom/dto/CheckInDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 import javax.validation.constraints.NotNull;
 
 /**
- * 签到请求DTO
+ * 签到请求DTO - 对齐数据库表结构
  * 负责人：严雯琪
  */
 @Data
@@ -16,5 +16,5 @@ public class CheckInDTO {
 
     @ApiModelProperty(value = "预约ID", required = true)
     @NotNull(message = "预约ID不能为空")
-    private Long reservationId;
+    private Integer reservationId;
 }
